LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-multiple-violations-01-expected.txt



PASS Violation report status OK.
FAIL Test number of sent reports. assert_equals: Report count was not what was expected. expected 2 but got 4
